requirements.txt
fastapi==0.109.1
uvicorn==0.27.0
python-jose[cryptography]==3.4.0
python-multipart==0.0.9
httpx==0.26.0
pydantic==2.5.3
pydantic-settings==2.1.0
python-dotenv==1.0.0
